size of

+++ 0x00-hello_world/6-size.c
@@ -15,10 +15,10 @@ int main(void)
 	long long int e;
 
 
-	printf("Size of an integer: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of characters: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of long integer: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of long long integer: %lu byte(s)", (unsigned long)sizeof(e));
+	printf("Size of an integer: %lu byte(s)\n", sizeof(a));
+	printf("Size of a float: %lu byte(s)\n", sizeof(b));
+	printf("Size of characters: %lu byte(s)\n", sizeof(c));
+	printf("Size of long integer: %lu byte(s)\n", sizeof(d));
+	printf("Size of long long integer: %lu byte(s)", sizeof(e));
 	return (0);
 }
